+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Recommended Gamevault Server Version: `v13.1.2`
 
 - Sync the currently played game with the discord Presence API (+)
 - Sync installed Gamevault games with Steam shortcuts (+)
+- Bug fix: Added increasing tick rate to download retry timer to avoid continuous spam
 - Bug fix: Media slider video was sometimes rendered on top of a popup
 - Bug fix: Crash on Media Slider navigation
 - Bug fix: Crash on create shortcut after installation